requirements.txt
Flask==3.0.0
google-cloud-datastore==2.15.1
python-dotenv==1.1.1
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0